# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Bitácora SOC API
  description: API completa para el sistema de Bitácora SOC con gestión de entradas, checklist de turno, usuarios y reportes
  version: 1.0.0
  contact:
    name: Bitácora SOC Team
servers:
  - url: http://192.168.100.50:3000/api
    description: Servidor de desarrollo (IP)
  - url: http://localhost:3000/api
    description: Servidor local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, user, guest]
        isActive:
          type: boolean
        theme:
          type: string
          enum: [light, dark, sepia, pastel]
        guestExpiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /users:
    get:
      tags:
        - Usuarios
      summary: Listar usuarios
      description: Obtiene la lista de todos los usuarios (solo admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    
    post:
      tags:
        - Usuarios
      summary: Crear usuario
      description: Crea un nuevo usuario (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - fullName
                - role
              properties:
                username:
                  type: string
                  example: jperez
                email:
                  type: string
                  example: jperez@example.com
                password:
                  type: string
                  example: password123
                fullName:
                  type: string
                  example: Juan Pérez
                role:
                  type: string
                  enum: [admin, user, guest]
                  example: user
      responses:
        '201':
          description: Usuario creado
        '400':
          description: Error de validación
  /users/me:
    get:
      tags:
        - Usuarios
      summary: Obtener perfil
      description: Obtiene el perfil del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    
    put:
      tags:
        - Usuarios
      summary: Actualizar perfil
      description: Actualiza el perfil del usuario autenticado
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                fullName:
                  type: string
                theme:
                  type: string
                  enum: [light, dark, sepia, pastel]
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Perfil actualizado
...
  /smtp:
    get:
      tags:
        - SMTP
      summary: Obtener configuración SMTP
      description: Obtiene la configuración del servidor de correo (solo admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configuración SMTP
    
    post:
      tags:
        - SMTP
      summary: Configurar SMTP
      description: Crea o actualiza la configuración SMTP (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
                - username
                - password
                - host
                - port
                - useTLS
                - senderName
                - senderEmail
                - recipients
                - sendOnlyIfRed
              properties:
                provider:
                  type: string
                  enum: [office365, aws-ses, elastic-email, google-mail, google-workspace, mailgun, custom]
                username:
                  type: string
                password:
                  type: string
                host:
                  type: string
                port:
                  type: integer
                useTLS:
                  type: boolean
                senderName:
                  type: string
                senderEmail:
                  type: string
                recipients:
                  type: array
                  items:
                    type: string
                sendOnlyIfRed:
                  type: boolean
      responses:
        '200':
          description: Configuración guardada
  /smtp/test:
    post:
      tags:
        - SMTP
      summary: Probar configuración SMTP
      description: Envía un correo de prueba (solo admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Correo enviado exitosamente
        '500':
          description: Error al enviar correo
...